refactor(): TwitchAPI
