+++ modules/twitch_api/utils.py
@@ -1,0 +1,7 @@
+import logging
+
+logger = logging.getLogger("uvicorn.error.twitch_api")
+
+def format_time(time_str):
+    """Format a timestamp into a readable time format."""
+    return time_str.strftime("%H:%M:%S")
